ap2-integration/pyproject.toml
[project]
name = "pokemon-ap2-integration"
version = "0.1.0"
description = "AP2 (Agent Payments Protocol) integration for Pokemon marketplace"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
]

[dependency-groups]
dev = []
